reference/stage-ultimate/application/requirements.txt
# Core Flask
flask==3.0.*
gunicorn==21.*

# Database
flask-sqlalchemy==3.1.*
psycopg2-binary==2.9.*

# Azure Key Vault
azure-identity==1.15.*
azure-keyvault-secrets==4.8.*

# Security
flask-wtf==1.2.*
